requirements.txt
django >3.2, <3.3
Pillow
djangorestframework
markdown 
django-filter